requirements.txt
annotated-types==0.7.0
anyio==4.9.0
boto3==1.38.42
botocore==1.38.42
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
construct==2.10.70
contourpy==1.3.2
cycler==0.12.1
exceptiongroup==1.3.0
fastapi==0.115.0
fonttools==4.58.4
h11==0.16.0
h5py==3.11.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
jmespath==1.0.1
kiwisolver==1.4.8
matplotlib==3.10.3
natsort==8.4.0
numpy==1.26.4
oct_converter==0.6.2
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
pydantic==2.11.7
pydantic_core==2.33.2
pydicom==2.4.3
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.10
pyvista==0.44.0
requests==2.32.4
s3transfer==0.13.0
scipy==1.11.4
scooby==0.10.1
six==1.17.0
sniffio==1.3.1
starlette==0.38.6
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.30.6
vtk==9.4.
gdcm==3.0.24
opencv-python==4.11.0.86
numpy==2.3.1
scipy==1.16.0
matplotlib==3.10.3
pydicom==3.0.1
